[chore] move Moim package to config
